Refactored menus into sub component and setup some styles
