+++ sass/layout/_sidebar-left.scss
@@ -257,7 +257,7 @@
         outline-style: none;
     }
 
-    .SidebarChannelMenu {
+    .SidebarMenu {
         height: 30px;
         opacity: 1;
 
@@ -300,7 +300,7 @@
         }
     }
 
-    .SidebarChannelMenu_menuButton {
+    .SidebarMenu_menuButton {
         padding: 0;
         z-index: 1;
         background: transparent;
@@ -308,18 +308,15 @@
         color: var(--sidebar-text);
         font-size: 16px;
         line-height: 16px;
-        height: 32px;
-        width: 32px;
         border-radius: 4px;
-        margin-top: -7px;
-        margin-right: 11px;
     }
-    .SidebarChannelMenu_menuButton:hover, .SidebarChannelMenu_menuButton:active, .SidebarChannelMenu_menuButton:focus {
+
+    .SidebarMenu_menuButton:hover, .SidebarMenu_menuButton:active, .SidebarMenu_menuButton:focus {
         cursor: pointer;
         background: rgba(255, 255, 255, 0.08);
         color: var(--sidebar-text);
     }
-    .SidebarChannelMenu_menuButton:focus {
+    .SidebarMenu_menuButton:focus {
         outline-style: none;
     }
 
@@ -437,7 +434,7 @@
         width: 100%;
         align-items: center;
 
-        > div {
+        > div:not(.SidebarMenu) {
             flex: 1 1 auto;
         }
 
@@ -447,6 +444,12 @@
             &.hide-arrow {
                 visibility: hidden;
             }
+        }
+
+        .SidebarMenu_menuButton {
+            height: 28px;
+            width: 28px;
+            margin-right: 16px;
         }
     }
 
@@ -548,6 +551,13 @@
                 @include opacity(1);
             }
         }
+
+        .SidebarMenu_menuButton {
+            height: 32px;
+            width: 32px;
+            margin-top: -7px;
+            margin-right: 11px;
+        }
     }
 
     .SidebarChannel:hover {
@@ -597,7 +607,7 @@
             font-weight: 300;
         }
 
-        .SidebarChannelMenu:not(.menuOpen) {
+        .SidebarMenu:not(.menuOpen) {
             display: none;
         }
     }
@@ -609,7 +619,7 @@
     .SidebarChannel .SidebarLink:hover {
         background: rgba(255, 255, 255, 0.08);
 
-        .SidebarChannelMenu {
+        .SidebarMenu {
             display: block;
         }
     }
